+++ changelog.md
@@ -2,9 +2,10 @@
 
 ### Next
 - Add forum notifications
-- Fix some filter sections to use OR instead of AND
+- Add notes and hidden from status lists to edit entry
 - Update missing news categories and fix filtering logic
 - Update staffMedia UI and add quick edit
+- Fix some filter sections to use OR instead of AND
 
 ### 0.40
 - Add long press images to view expanded